Cleanup
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@1854 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ common/JackAPI.cpp
@@ -239,7 +239,7 @@ static inline void WaitGraphChange()
 	
     if (manager && control && manager->IsPendingChange()) {
         JackLog("WaitGraphChange...\n");
-        JackSleep(control->fPeriodUsecs * 1.1f);
+        JackSleep(int(control->fPeriodUsecs * 1.1f));
     }
 }
